+++ CHANGELOG.md
@@ -7,7 +7,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## UNRELEASED
 
-_Nothing yet.._
+### Breaking
+
+- Updated [`eslint`]() to `7.11.0`
+- Updated [`eslint-plugin-jest`](https://github.com/jest-community/eslint-plugin-jest/releases) to `24.1.0`
+- Updated [`eslint-plugin-react-hooks`](https://github.com/facebook/react/blob/master/packages/eslint-plugin-react-hooks/CHANGELOG.md) to `4.1.2`
+- Updated [`prettier`](https://github.com/prettier/prettier/releases) to `2.1.2`
+
+### Changed
+
+- Turned off `backpack/use-components` due to a bug with `autoImport` not correctly turning off when configured.
+
+- Updated [`eslint-config-airbnb`](https://github.com/airbnb/javascript/blob/master/packages/eslint-config-airbnb/CHANGELOG.md) to `18.2.0`
+- Updated [`eslint-plugin-jsx-a11y`](https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/blob/master/CHANGELOG.md) to `6.3.1`
+- Updated [`eslint-plugin-react`]() to `7.21.4`
 
 ## 7.0.0 - Upgraded `eslint-plugin-backpack`